Moved temperature conversion to it's own folder
Fixes #5
